1)Добавляет view для пустого списка точек 2)Добавляет view для контейнера точек
