+++ src/main.js
@@ -3,7 +3,7 @@ import PointModel from './model/travel-route-model.js';
 import { getPoints, getDestinations, getOffers, getOffersByType } from './mock/travel-route-mock.js';
 
 const filtersPosition = document.querySelector('.trip-controls__filters');
-const containerPosition = document.querySelector('.trip-events');
+const contentContainer = document.querySelector('.trip-events');
 
 const points = getPoints();
 const destinations = getDestinations();
@@ -11,5 +11,5 @@ const offers = getOffers();
 const offersByType = getOffersByType();
 
 const pointModel = new PointModel({points, destinations, offers, offersByType});
-const travelRoutePresenter = new TravelRoutePresenter(filtersPosition, containerPosition, pointModel);
+const travelRoutePresenter = new TravelRoutePresenter(filtersPosition, contentContainer, pointModel);
 travelRoutePresenter.init();
