+++ src/view/filter.js
@@ -6,7 +6,7 @@ function createFilterTemplate() {
               <h2 class="visually-hidden">Filter events</h2>
               <form class="trip-filters" action="#" method="get">
                 <div class="trip-filters__filter">
-                  <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
+                  <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
                   <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
                 </div>
 
@@ -16,7 +16,7 @@ function createFilterTemplate() {
                 </div>
 
                 <div class="trip-filters__filter">
-                  <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past" checked>
+                  <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
                   <label class="trip-filters__filter-label" for="filter-past">Past</label>
                 </div>
