+++ src/view/point-container.js
@@ -1,0 +1,12 @@
+import AbstractView from '../framework/view/abstract-view.js';
+
+function createContainer() {
+  return (`
+  <ul class="trip-events__list"></ul>`);
+}
+export default class PointContainerView extends AbstractView {
+
+  get template() {
+    return createContainer();
+  }
+}
